+++ db/migrate/20130410170251_add_mrn_organization_to_patients.rb
@@ -1,0 +1,5 @@
+class AddMrnOrganizationToPatients < ActiveRecord::Migration
+  def change
+    add_column :patients, :mrn_organization, :string
+  end
+end