+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.9.0
 
 ### Enhancements
+- MRN organization added to patients
 - Prescreens inline editing has been removed
 - **Gem Changes**
   - Updated to Rails 4.0.0.beta1